% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcal.R
\name{calculateKcalPerModel}
\alias{calculateKcalPerModel}
\title{Calculate kcal per model}
\usage{
calculateKcalPerModel(model.res, kcal.params = c(carbs = 4, prot = 4,
  satu = 9, mono = 9, poly = 9))
}
\arguments{
\item{model.res}{model results i.e. \code{calculateParams()} function output}

\item{kcal.params}{vector with kcal values for carbs (kcal for carbohydrates), 
prot (kcal for protein), satu (kcal for saturated fat),
mono (kcal for monosaturated fat), poly (kcal for polyunsaturated fat)
(default: \code{c(carbs = 4, prot  = 4, satu  = 9, mono  = 9, poly  = 9)})}
}
\value{
matrix with model results and additional column kcal
}
\description{
Calculate kcal per model
}
\author{
Konrad J. Debski
}
