% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateParams}
\alias{calculateParams}
\title{Calculate params}
\usage{
calculateParams(params.matrix, model.conf, range = c(-0.1, 0),
  only.in.range = FALSE, add.kcal = TRUE, kcal.params = c(carbs = 4,
  prot = 4, satu = 9, mono = 9, poly = 9))
}
\arguments{
\item{params.matrix}{matrix with params to test (output from \code{fsParams::generateParams()}}

\item{model.conf}{list with model configuration (params)}

\item{range}{two element vector (default: \code{c(-0.1, 0)}}

\item{only.in.range}{logical. If \code{TRUE} returns only results in range}

\item{add.kcal}{logical. If \code{TRUE} kcal column is added.}

\item{kcal.params}{vector with kcal values for carbs (kcal for carbohydrates), 
prot (kcal for protein), satu (kcal for saturated fat),
mono (kcal for monosaturated fat), poly (kcal for polyunsaturated fat).
Ignored if \code{add.kcal = FALSE}.
(default: \code{c(carbs = 4, prot  = 4, satu  = 9, mono  = 9, poly  = 9)})}
}
\value{
matrix with model, carbs, prot, satu, mono, poly, is_in_range and kcal 
(if \code{add.kcal = TRUE})
}
\description{
Calculate params
}
