% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{doCalculations}
\alias{doCalculations}
\title{Do model calculations}
\usage{
doCalculations(r, range, shifts, x)
}
\arguments{
\item{r}{single column data.frame "R"}

\item{range}{two element vector}

\item{shifts}{vector with carbs.shift (shift for carbohydrates), 
prot.shift (shift for protein), satu.shift (shift for saturated fat), 
mono.shift (shift for monosaturated fat), poly.shift (shift for polyunsaturated fat)}

\item{x}{data.frame. Whole data i.e. usda_nutrients}
}
\value{
vector with rsq and shifts
}
\description{
Do model calculations
}
\author{
Konrad J. Debski
}
