% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generateModel}
\alias{generateModel}
\title{Generate model for data}
\usage{
generateModel(data)
}
\arguments{
\item{data}{data frame for model with 'R', 'carbs', 'prot', 
'satu', 'mono', 'poly' columns}
}
\value{
gam model
}
\description{
Model formula: \code{R~carbs+prot+satu+mono+poly}
}
\author{
Konrad J. Debski
}
