% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getData}
\alias{getData}
\title{Get data}
\usage{
getData(x, range, shift, column)
}
\arguments{
\item{x}{data.frame with data}

\item{range}{vector length of 2 c(from, to)}

\item{shift}{vector length of 1}

\item{column}{integer column id}
}
\value{
data.frame data subset
}
\description{
Get data
}
\author{
Konrad J. Debski
}
