% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getModelRSq}
\alias{getModelRSq}
\title{Get model R SQ}
\usage{
getModelRSq(model)
}
\arguments{
\item{model}{gam model (output from gam function from mgcv package)}
}
\value{
model R SQ
}
\description{
Get model R SQ
}
\author{
Konrad J. Debski
}
