% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getRange4Period}
\alias{getRange4Period}
\title{Get range for period}
\usage{
getRange4Period(x, period)
}
\arguments{
\item{x}{data.frame}

\item{period}{vector length of one. Period id e.g. 1, 2, 3 or 4}
}
\value{
range
}
\description{
Get range for period
}
\author{
Konrad J. Debski
}
