% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{makeCalculations}
\alias{makeCalculations}
\title{Make model calculations in paraller}
\usage{
makeCalculations(x, period, shifts, cores = 6)
}
\arguments{
\item{x}{data.frame. Whole data i.e. usda_nutrients}

\item{period}{period id e.g. 1, 2, 3, 4}

\item{shifts}{list with}

\item{cores}{number of cores to use for calculation}
}
\value{
matrix with results
}
\description{
Make model calculations in paraller
}
\author{
Konrad J. Debski
}
