% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_configuration}
\alias{model_configuration}
\title{Model configuration for periods}
\format{list of lists with model configurations for each period 
intercept, carbs, prot, satu, mono, poly}
\usage{
data(model_configuration)
}
\description{
Model configuration for periods
}
\examples{
data(model_configuration)
}
\keyword{datasets}
