% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params}
\alias{params}
\title{Precalculated params}
\format{matrix with carbs, prot, satu, mono, poly}
\source{
\code{params <- generateParams(params_conf)}
}
\usage{
data(params)
}
\description{
Precalculated params
}
\examples{
data(params)
}
\keyword{datasets}
