% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_conf}
\alias{params_conf}
\title{Params configuration used in paper}
\format{A list with ranges and steps for carbs, prot, satu, mono, poly}
\usage{
data(params_conf)
}
\description{
Params configuration used in paper
}
\examples{
data(params_conf)
}
\keyword{datasets}
