% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data for model}
\usage{
prepareData(x, r, range, shifts)
}
\arguments{
\item{x}{data.frame with data "Carbohydrates", "Protein", "Saturated_Fat", 
"Monounsaturated_Fat", "Polyunsaturated_Fat"}

\item{r}{single column data.frame "R"}

\item{range}{two element vector}

\item{shifts}{vector with carbs.shift (shift for carbohydrates), 
prot.shift (shift for protein), satu.shift (shift for saturated fat), 
mono.shift (shift for monosaturated fat), poly.shift (shift for polyunsaturated fat)}
}
\value{
data.frame with r, carbs, prot, satu, mono, poly
}
\description{
Prepare data for model
}
\author{
Konrad J. Debski
}
