% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{prepareModelConfMatrix}
\alias{prepareModelConfMatrix}
\title{Prepare model configuration matrix}
\usage{
prepareModelConfMatrix(model.conf, params.matrix)
}
\arguments{
\item{model.conf}{list model configuration}

\item{params.matrix}{matrix with params}
}
\value{
matrix with model configuration
}
\description{
Prepare model configuration matrix
}
\author{
Konrad J. Debski
}
