% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{prepareShifts}
\alias{prepareShifts}
\title{Prepare data shifts}
\usage{
prepareShifts(init.range = 0:20)
}
\arguments{
\item{init.range}{vector of integers to generate 
all shifts combinations (default: \code{0:20})}
}
\value{
list of named vectors with data shifts 
(carbs.shift, prot.shift, satu.shift, mono.shift, poly.shift)
}
\description{
Prepare data shifts
}
\author{
Konrad J. Debski
}
