% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shifts}
\alias{shifts}
\title{Precalculated shifts for model}
\format{A list of 4084101 data frames with columns carbs.shift, prot.shift, satu.shift, 
mono.shift, poly.shift}
\source{
\code{shifts <- prepareShifts(0:20)}
}
\usage{
data(shifts)
}
\description{
Precalculated shifts for model
}
\examples{
data(shifts)
}
\keyword{datasets}
