% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usda_nutrients}
\alias{usda_nutrients}
\title{Table of nutrients availability from USDA/Center for Nutrition Policy and Promotion}
\format{A data.frame object.}
\source{
\href{https://www.cnpp.usda.gov/tools/IFS/Query.htm}{Nutrient Content of the U.S. Food Supply}
}
\usage{
data(usda_nutrients)
}
\description{
Table of nutrients availability from USDA/Center for Nutrition Policy and Promotion
}
\examples{
data(usda_nutrients)
}
\references{
Food and Nutrion Service U.S. Department of Agriculture
(\href{https://www.fns.usda.gov/}{Food and Nutrion Service U.S. Department of Agriculture})
}
\keyword{datasets}
